[#14]refactor: exception, service 폴더 생성
