+++ app/route/advice/advice_router.py
@@ -4,7 +4,7 @@ from starlette.responses import JSONResponse
 from app.route.advice.models.correct_request import CorrectRequest
 from app.route.advice.models.hint_request import HintRequest
 from app.web.models.success_reponse import SuccessResponse
-from app.route.advice import ai_service
+from app.route.advice.service import ai_service
 
 advice_router = APIRouter()
